Rename test files and restructure test directory to make it clearer what is being tested and where. Include new unit test for Express routes and update npm test script.
