+++ package.json
@@ -8,7 +8,7 @@
   },
   "scripts": {
     "start": "nodemon --exec babel-node bin/www",
-    "test": "./node_modules/.bin/mocha --reporter spec ./test/*.js; testcafe chrome test/feature/* --app \"node bin/www\"",
+    "test": "./node_modules/.bin/mocha --reporter spec ./test/*.js; ./node_modules/.bin/mocha --reporter spec ./test/routes/*.js; testcafe chrome test/feature/* --app \"node bin/www\"",
     "jshint": "./node_modules/.bin/jshint",
     "coverage": "./node_modules/.bin/istanbul cover ./node_modules/.bin/_mocha",
     "coveralls": "npm run coverage -- --report lcovonly && cat ./coverage/lcov.info | coveralls"
